requirements.txt
alembic==1.5.8
autopep8==1.5.7
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.14.5
click==7.1.2
cloudinary==1.28.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
Flask-WTF==1.0.0
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-http-client==3.3.4
sendgrid==6.9.3
six==1.15.0
SQLAlchemy==1.4.12
starkbank-ecdsa==2.0.3
toml==0.10.2
urllib3==1.26.7
Werkzeug==1.0.1
WTForms==3.0.0
